refactor: move user module to core module
