+++ src/core/user/entities/user.entity.ts
