+++ src/core/user/entities/user.model.ts
